refactor(workflow): peel eval-defense + structured forcing into 2 mixins
workflow.py (1269 ln) welded a generic async engine to two application-
specific "forced delivery" stacks. Relocate them verbatim (a move, not a
merge — the 5 runners share only the build->track->run->swallow frame;
their middles genuinely differ):

- workflow_agents.py (WorkflowAgentsMixin): the enforcement wind-down +
  commit-first stack (_run_enforced_agent, _synthesize_dead_scout,
  draft_findings, _draft_findings_with_lease, _scout_ledger,
  _emit_dead_scout_synthesis, _configure_session_enforcement,
  _session_messages, _emit_commitment_terminus).
- workflow_structured.py (WorkflowStructuredMixin): the schema= forcing
  stack (_run_structured_agent, _forced_structured_commit,
  _carry_exploration) + the shared module helpers _STRUCTURED_*,
  _named_tool_choice, _schema_satisfied.

workflow.py keeps only the engine (agent dispatch, _run_agent,
_run_session_turn, budget/lease/semaphore, parallel/pipeline, timeout,
observability) and composes WorkflowContext(WorkflowAgentsMixin,
WorkflowStructuredMixin): 1269 -> 730 ln (<800). Public surface preserved
via the compose class + a _schema_satisfied re-export (a test imports it
by name). The one compose-cycle edge (draft_findings catches the core
WorkflowBudgetExceeded) is broken with a function-local import; the one
cross-mixin helper (_named_tool_choice, used by both forced-delivery
stacks) lives in workflow_structured and is imported by workflow_agents.

Pure relocation, no behavior change. pytest 1368 green, ruff +
boundaries (application/domain/sdk) green.
