+++ opencollab/opencollab/application/workflow.py
@@ -48,14 +48,12 @@ from opencollab.application.ports import (
     WorkingTreeProbe,
 )
 from opencollab.application.session_run import DEFAULT_COMMIT_RESERVE, ENFORCEMENT_OFF
-from opencollab.application.structured_output import StructuredOutputTool
-from opencollab.application.submit_findings import (
-    SUBMIT_TOOL_NAME,
-    SubmitFindingsTool,
-    build_dead_scout_synthesis_prompt,
-    commitment_terminus_payload,
-    format_findings_report,
-    harvest_findings,
+from opencollab.application.workflow_agents import WorkflowAgentsMixin
+from opencollab.application.workflow_structured import (
+    WorkflowStructuredMixin,
+)
+from opencollab.application.workflow_structured import (
+    _schema_satisfied as _schema_satisfied,
 )
 
 logger = logging.getLogger(__name__)
@@ -70,42 +68,6 @@ DEFAULT_MAX_CONCURRENCY = 4
 DEFAULT_DEADLINE_MARGIN_SECONDS = 120.0
 DEFAULT_INTERNAL_COMMIT_TIMEOUT_SECONDS = 120.0
 
-# Appended to a schema= prompt: the agent must finish by emitting structured
-# output via the injected tool rather than free-text.
-_STRUCTURED_INSTRUCTION = (
-    "\n\nFinish by calling the `structured_output` tool — do not answer in "
-    "free text."
-)
-
-# Corrective message that seeds the forced-commit retry session when the first
-# run produced no valid structured payload. The retry session is restricted to
-# the single capture tool with a named-function ``tool_choice`` (force exactly
-# ``structured_output``) — graceful, not guaranteed: an endpoint may 400-reject
-# the forced choice and degrade to ``auto``, after which the model can still
-# answer in prose. This prompt leads with an explicit MUST-call / no-prose
-# imperative and tells it to commit its final result NOW from what it already
-# gathered rather than answer in free text.
-_STRUCTURED_RETRY = (
-    "You MUST call the `structured_output` tool now, exactly once, with your "
-    "final result based on what you have already gathered, conforming to the "
-    "required schema. Do not explore further or answer in prose."
-)
-
-
-def _named_tool_choice(tool_name: str) -> dict[str, Any]:
-    """OpenAI-style named-function ``tool_choice`` forcing exactly ``tool_name``.
-
-    More precise than the bare ``"required"`` (force *some* tool) string: it
-    names the single tool the corrective turn must call. Stricter
-    OpenAI-compatible endpoints (observed: DashScope 400-rejects a bare
-    ``"required"`` for several repos and silently degrades to ``auto``) are more
-    likely to honour this explicit dict. It rides through the LLM stack
-    unchanged (the OpenAI SDK accepts a dict ``tool_choice``); if an endpoint
-    still rejects it, ``SessionRunUseCase._complete`` degrades it ONCE to
-    ``"auto"`` on a 400 exactly as it does for ``"required"`` today.
-    """
-    return {"type": "function", "function": {"name": tool_name}}
-
 # Per-agent token budget handed to a session when the workflow budget is
 # unbounded (``budget_total is None``). The session still needs a finite cap.
 UNBOUNDED_SESSION_BUDGET = 1_000_000
@@ -115,25 +77,6 @@ Thunk = Callable[[], Awaitable[Any]]
 
 # A pipeline stage receives (previous result, original item, item index).
 Stage = Callable[[Any, Any, int], Awaitable[Any]]
-
-
-def _schema_satisfied(captured: Any, schema: dict[str, Any]) -> bool:
-    """Minimal acceptance check for a captured structured payload.
-
-    The ``StructuredOutputTool`` already validated the payload against the full
-    schema before storing it, so this is light hardening, not a re-validation:
-    it rejects a missing capture and a dict that omits any of the schema's
-    required top-level keys (e.g. an empty ``{}`` that slipped through), which
-    are treated like a miss so the forced corrective turn runs.
-    """
-    if not isinstance(captured, dict):
-        return False
-    required = schema.get("required") if isinstance(schema, dict) else None
-    if not required:
-        # No required keys: the tool already validated the payload, so any
-        # captured dict (even ``{}``) is an accepted commit, not a miss.
-        return True
-    return all(key in captured for key in required)
 
 
 class WorkflowBudgetExceeded(Exception):
@@ -202,7 +145,7 @@ class _BudgetLease:
         return max(0, self.total - spent)
 
 
-class WorkflowContext:
+class WorkflowContext(WorkflowAgentsMixin, WorkflowStructuredMixin):
     """Primitives a workflow function uses to orchestrate agent sessions."""
 
     def __init__(
@@ -625,465 +568,6 @@ class WorkflowContext:
             await self.log(f"agent failed ({label or 'agent'}): {exc}")
             return None
 
-    async def _run_enforced_agent(
-        self,
-        prompt: str,
-        *,
-        label: str | None,
-        tools: Sequence[Any] | None,
-        isolation: bool,
-        tool_choice: str | None,
-        thinking: bool | None,
-        timeout: float | None,
-        budget: int | None,
-        enforcement_strength: str,
-        commit_reserve: int,
-        harvest_fallback: str | None = None,
-    ) -> str | None:
-        """Run a scout under the enforcement wind-down (STEP 0).
-
-        Mirrors ``_run_agent`` but injects a ``submit_findings`` capture tool, arms
-        the runner's structural commit brake, and HARVESTS a usable report: the
-        captured payload if present, else the final text, else a "(partial …)"
-        salvage from the transcript — so a chopped scout never yields a bare
-        "(scout died)". A successful capture sets the cancel event so the loop halts
-        at once (commit-first friendly), exactly as the structured path does. Emits
-        one ``commitment_terminus`` metric per scout to the orchestration trace.
-        """
-        deadline = self._timeout_deadline(timeout)
-        session_budget = self._capped_session_budget(budget)
-        capture_done = asyncio.Event()
-        submit_tool = SubmitFindingsTool(on_capture=capture_done.set)
-        combined_tools = [*(tools or []), submit_tool]
-        try:
-            session = self._factory.build_workflow_session(
-                prompt=prompt,
-                budget=session_budget,
-                tools=combined_tools,
-                isolation=isolation,
-                label=label,
-                tool_choice=tool_choice,
-                thinking=thinking,
-            )
-        except Exception as exc:  # noqa: BLE001 — factory failure must not abort the fleet
-            await self.log(f"agent build failed ({label or 'agent'}): {exc}")
-            return None
-
-        self._track_session(session)
-        self._configure_session_enforcement(
-            session, enforcement_strength, commit_reserve
-        )
-        text: str | None = None
-        try:
-            text = await self._run_session_turn(
-                session,
-                prompt,
-                deadline=deadline,
-                cancel_event=capture_done,
-            )
-        except CallerTimeoutError:
-            await self.log(f"agent timed out ({label or 'agent'}) after {timeout}s")
-        except Exception as exc:  # noqa: BLE001 — one dead agent never kills the fleet
-            await self.log(f"agent failed ({label or 'agent'}): {exc}")
-
-        # Harvest is the backstop even on a timeout/exception: whatever the scout
-        # already gathered (captured payload, prose, or the harness-authored
-        # evidence ledger) is salvaged — never a bare "(scout died)".
-        ledger = self._scout_ledger(session)
-        report = harvest_findings(
-            submit_tool.captured, text or "", self._session_messages(session), ledger=ledger,
-            draft=harvest_fallback,
-        )
-        # STEP 2 (rare-case backstop): a DEAD scout — no structured commit
-        # (``captured is None``) yet a non-empty ledger of what it gathered —
-        # triggers ONE bounded transcript-only synthesizer call (submit_findings
-        # only, forced, cite-or-abstain). With STEP 0's wind-down live this fires
-        # seldom (scouts are force-committed at ~80%); it salvages the chopped /
-        # errored / strayed tail. Gated by construction: this method only runs when
-        # enforcement is on.
-        if (
-            submit_tool.captured is None
-            and ledger
-            and not self._active_call_has_pending_cleanup()
-        ):
-            synthesized = await self._synthesize_dead_scout(
-                session, label, commit_reserve=commit_reserve
-            )
-            if synthesized and synthesized.strip():
-                report = synthesized
-        self._emit_commitment_terminus(session, label, submit_tool, report)
-        return report if report else text
-
-    async def _synthesize_dead_scout(
-        self, dead_session: Any, label: str | None, *, commit_reserve: int
-    ) -> str | None:
-        """Salvage a dead/empty scout with ONE bounded transcript-only LLM call.
-
-        Its ONLY input is the scout's harness-authored evidence ledger + raw tool
-        results; its ONLY tool is ``submit_findings`` with a forced (named-function)
-        ``tool_choice`` and the cite-or-abstain post-validation — NO exploration
-        tools, so the salvage cannot wander or fabricate. Returns the formatted
-        findings (or a valid ``insufficient_evidence`` abstention) on a successful
-        capture, else ``None`` so the caller keeps the harvested partial. Bounded by
-        ``commit_reserve`` (the reserve sized for a single submit turn) and clamped
-        to the live global remaining.
-        """
-        ledger = self._scout_ledger(dead_session)
-        messages = self._session_messages(dead_session)
-        prompt = build_dead_scout_synthesis_prompt(ledger, messages)
-        capture_done = asyncio.Event()
-        submit_tool = SubmitFindingsTool(on_capture=capture_done.set)
-        synth_label = f"{label}:synth" if label else "synth"
-        session_budget = self._capped_session_budget(commit_reserve)
-        try:
-            session = self._factory.build_workflow_session(
-                prompt=prompt,
-                budget=session_budget,
-                tools=[submit_tool],
-                isolation=False,
-                label=synth_label,
-                tool_choice=_named_tool_choice(SUBMIT_TOOL_NAME),
-                thinking=False,
-            )
-        except Exception as exc:  # noqa: BLE001 — a failed salvage must not abort the fleet
-            await self.log(f"dead-scout synth build failed ({synth_label}): {exc}")
-            return None
-
-        self._track_session(session)
-        try:
-            deadline = self._internal_commit_deadline()
-            await self._run_session_turn(
-                session,
-                prompt,
-                deadline=deadline,
-                cancel_event=capture_done,
-            )
-        except Exception as exc:  # noqa: BLE001 — one dead salvage never kills the fleet
-            await self.log(f"dead-scout synth failed ({synth_label}): {exc}")
-            return None
-
-        captured = submit_tool.captured
-        report = format_findings_report(captured) if captured is not None else ""
-        self._emit_dead_scout_synthesis(synth_label, ledger, captured, bool(report.strip()))
-        return report if report.strip() else None
-
-    async def draft_findings(
-        self, prompt: str, *, label: str | None = None, budget: int | None = None
-    ) -> dict[str, Any] | None:
-        """STEP 5b commit-first: ONE bounded submit-only call that commits a
-        structured ``submit_findings`` DRAFT from STATIC context (the pre-recon fact
-        sheet) BEFORE any exploration, returning the captured payload (or ``None``).
-
-        Reuses the validated dead-scout-synth wiring exactly — ``tools=[submit_findings]``
-        only, a named-function (forced) ``tool_choice``, ``thinking=False`` — so the
-        draft cannot wander or fabricate and the call is a single constrained turn.
-        It touches NO part of the session FSM: the exploring scout that consumes this
-        draft runs the unchanged capture→cancel→harvest path. Cost is one bounded
-        call per scout, clamped to ``budget`` (sized to ``commit_reserve``) and to the
-        live global remaining. Skips gracefully (``None``) if the shared pool is spent
-        or the factory/session errors, so a failed draft never aborts the fleet.
-        """
-        call_task = asyncio.current_task()
-        if call_task is not None:
-            self._active_call_tasks.add(call_task)
-        slot_acquired = False
-        slot_handed_to_cleanup = False
-        try:
-            await self._semaphore.acquire()
-            slot_acquired = True
-            try:
-                lease = await self._acquire_budget_lease(budget, over_budget_ok=False)
-            except WorkflowBudgetExceeded:
-                return None
-            token = self._active_budget_lease.set(lease)
-            try:
-                return await self._draft_findings_with_lease(
-                    prompt,
-                    label=label,
-                    budget=budget,
-                )
-            finally:
-                self._active_budget_lease.reset(token)
-                slot_handed_to_cleanup = self._release_lease_when_quiescent(lease)
-        finally:
-            if slot_acquired and not slot_handed_to_cleanup:
-                self._semaphore.release()
-            if call_task is not None:
-                self._active_call_tasks.discard(call_task)
-
-    async def _draft_findings_with_lease(
-        self, prompt: str, *, label: str | None, budget: int | None
-    ) -> dict[str, Any] | None:
-        session_budget = self._capped_session_budget(budget)
-        capture_done = asyncio.Event()
-        submit_tool = SubmitFindingsTool(on_capture=capture_done.set)
-        try:
-            session = self._factory.build_workflow_session(
-                prompt=prompt,
-                budget=session_budget,
-                tools=[submit_tool],
-                isolation=False,
-                label=label,
-                tool_choice=_named_tool_choice(SUBMIT_TOOL_NAME),
-                thinking=False,
-            )
-        except Exception as exc:  # noqa: BLE001 — a failed draft must not abort the fleet
-            await self.log(f"draft build failed ({label or 'draft'}): {exc}")
-            return None
-        self._track_session(session)
-        try:
-            deadline = self._internal_commit_deadline()
-            await self._run_session_turn(
-                session,
-                prompt,
-                deadline=deadline,
-                cancel_event=capture_done,
-            )
-        except Exception as exc:  # noqa: BLE001 — one dead draft never kills the fleet
-            await self.log(f"draft failed ({label or 'draft'}): {exc}")
-            return None
-        return submit_tool.captured
-
-    @staticmethod
-    def _scout_ledger(session: Any) -> list[dict[str, Any]]:
-        """The scout's harness-authored evidence ledger (STEP 2), or [] when a
-        duck-typed session/state does not carry one."""
-        state = getattr(session, "state", None)
-        turn = getattr(state, "turn", None)
-        ledger = getattr(turn, "scout_ledger", None)
-        return list(ledger) if ledger else []
-
-    def _emit_dead_scout_synthesis(
-        self, label: str | None, ledger: list[dict[str, Any]], captured: dict | None, salvaged: bool
-    ) -> None:
-        """Trace one ``dead_scout_synthesis`` event (no-op without a tracer) so the
-        rare salvage is auditable: how big the ledger was, whether a payload was
-        captured, and the anchor count of the salvaged findings."""
-        if self._tracer is None:
-            return
-        findings = (captured or {}).get("findings") or []
-        self._tracer.log_step(
-            step_type="dead_scout_synthesis",
-            payload={
-                "role": label,
-                "ledger_size": len(ledger),
-                "salvaged": salvaged,
-                "insufficient_evidence": bool((captured or {}).get("insufficient_evidence")),
-                "evidence_anchor_count": sum(
-                    1 for f in findings if str(f.get("evidence_anchor") or "").strip()
-                ),
-            },
-        )
-
-    @staticmethod
-    def _configure_session_enforcement(
-        session: Any,
-        enforcement_strength: str,
-        commit_reserve: int,
-    ) -> None:
-        """Arm the session runner's wind-down post-build (the agent already carries
-        the submit tool). Defensive: a duck-typed session without a configurable
-        runner is left as-is rather than aborting the scout."""
-        runner = getattr(session, "runner", None)
-        configure = getattr(runner, "configure_enforcement", None)
-        if callable(configure):
-            configure(
-                enforcement_strength=enforcement_strength,
-                commit_reserve=commit_reserve,
-            )
-
-    @staticmethod
-    def _session_messages(session: Any) -> list[dict[str, Any]]:
-        state = getattr(session, "state", None)
-        messages = getattr(state, "messages", None)
-        if messages is None:
-            messages = getattr(session, "messages", None)
-        return list(messages) if messages else []
-
-    def _emit_commitment_terminus(
-        self, session: Any, label: str | None, submit_tool: SubmitFindingsTool, report: str | None
-    ) -> None:
-        """Emit one ``commitment_terminus`` event per scout to orchestration.jsonl
-        (no-op when no tracer is wired)."""
-        if self._tracer is None:
-            return
-        state = getattr(session, "state", None)
-        payload = commitment_terminus_payload(
-            role=label,
-            captured=submit_tool.captured,
-            wind_down_done=bool(getattr(state, "wind_down_done", False)),
-            used_tokens=int(getattr(state, "used_tokens", 0) or 0),
-            max_budget_tokens=int(getattr(session, "max_budget_tokens", 0) or 0),
-            wind_down_token_mark=int(getattr(state, "wind_down_token_mark", 0) or 0),
-            artifact=report or "",
-        )
-        self._tracer.log_step(step_type="commitment_terminus", payload=payload)
-
-    async def _run_structured_agent(
-        self,
-        prompt: str,
-        *,
-        schema: dict[str, Any],
-        label: str | None,
-        tools: Sequence[Any] | None,
-        isolation: bool,
-        timeout: float | None = None,
-        budget: int | None = None,
-    ) -> dict | None:
-        """Run a schema-bound session, returning the validated payload or None.
-
-        First pass — free exploration: the full toolset
-        ``[capture_tool, *tools]`` is offered with ``tool_choice`` left at the
-        endpoint default so the agent can grep/read before committing, and it is
-        instructed to finish by calling ``structured_output``.
-
-        Corrective pass — forced commit: the trigger is a genuinely-empty
-        capture (``_schema_satisfied(captured)`` is False after the first pass),
-        NOT a free-text stop reason — so a markup-leaked tool call that the
-        parser already resolved into ``captured`` does NOT spuriously fire it.
-        When it fires, a second session restricted to ONLY the capture tool with
-        a named-function ``tool_choice`` (``_named_tool_choice``) is built,
-        seeded with the first pass's conversation (its exploration is copied
-        over) and an explicit 'you MUST call structured_output, do not answer in
-        prose' instruction, so it commits from what was actually gathered rather
-        than the bare prompt. This raises the odds of a commit but does NOT
-        guarantee one: some OpenAI-compatible endpoints (observed: DashScope)
-        400-reject a forced ``tool_choice`` and ``session_run`` degrades it once
-        to ``auto``, after which the model may still answer in prose — a
-        still-missing capture yields ``None``.
-
-        A successful capture sets a cancel event that the session's precheck
-        observes before each LLM call, halting the loop immediately. Without
-        it, a model that keeps re-calling structured_output after acceptance
-        burns the whole session budget (observed live: one valid capture
-        followed by 28 wasted calls until budget death).
-        """
-        deadline = self._timeout_deadline(timeout)
-        capture_done = asyncio.Event()
-        capture_tool = StructuredOutputTool(schema, on_capture=capture_done.set)
-        seeded_prompt = prompt + _STRUCTURED_INSTRUCTION
-        combined_tools = [capture_tool, *(tools or [])]
-        session_budget = self._capped_session_budget(budget)
-        try:
-            session = self._factory.build_workflow_session(
-                prompt=seeded_prompt,
-                budget=session_budget,
-                tools=combined_tools,
-                isolation=isolation,
-                label=label,
-                thinking=False,
-            )
-        except Exception as exc:  # noqa: BLE001 — factory failure must not abort the fleet
-            await self.log(f"structured agent build failed ({label or 'agent'}): {exc}")
-            return None
-
-        # Track immediately so tokens count even if a run_loop raises midway.
-        self._track_session(session)
-        try:
-            await self._run_session_turn(
-                session,
-                seeded_prompt,
-                deadline=deadline,
-                cancel_event=capture_done,
-            )
-            if _schema_satisfied(capture_tool.captured, schema):
-                return capture_tool.captured
-        except CallerTimeoutError:
-            await self.log(f"structured agent timed out ({label or 'agent'}) after {timeout}s")
-            return None
-        except Exception as exc:  # noqa: BLE001 — one dead agent never kills the fleet
-            await self.log(f"structured agent failed ({label or 'agent'}): {exc}")
-            return None
-
-        # Corrective pass (only when the capture is genuinely empty above): force
-        # the structured commit on a single-tool session pinned to a
-        # named-function ``tool_choice`` — graceful, NOT guaranteed (an endpoint
-        # may 400-reject it and degrade to ``auto``). Reusing the same
-        # capture_tool keeps ``captured`` and the cancel event live across both
-        # passes; the first session is handed in so its exploration history is
-        # carried into the corrective turn.
-        try:
-            retry_timeout = self._remaining_timeout(deadline)
-        except CallerTimeoutError:
-            await self.log(
-                f"structured agent timed out ({label or 'agent'}) after {timeout}s"
-            )
-            return None
-        return await self._forced_structured_commit(
-            prompt,
-            session,
-            capture_tool,
-            capture_done,
-            schema=schema,
-            label=label,
-            isolation=isolation,
-            timeout=retry_timeout,
-            budget=budget,
-        )
-
-    async def _forced_structured_commit(
-        self,
-        prompt: str,
-        prior_session: Any,
-        capture_tool: StructuredOutputTool,
-        capture_done: asyncio.Event,
-        *,
-        schema: dict[str, Any],
-        label: str | None,
-        isolation: bool,
-        timeout: float | None,
-        budget: int | None,
-    ) -> dict | None:
-        """Build a single-tool, forced-``tool_choice`` corrective session.
-
-        The session is pinned to a named-function ``tool_choice`` (force exactly
-        ``structured_output``) and seeded with an explicit 'you MUST call the
-        tool, do not answer in prose' instruction. This strongly pushes — but,
-        on endpoints that 400-reject a forced choice and degrade to ``auto``,
-        does not guarantee — a structured commit.
-
-        The first pass's conversation (its grep/file_read tool results and the
-        understanding the model built) is copied from ``prior_session`` into the
-        corrective session before the retry message is added, so the forced
-        commit fills the schema from real exploration rather than from the bare
-        prompt — without this carry-over the ``_STRUCTURED_RETRY`` instruction to
-        commit "based on what you have already gathered" would address a blank
-        session that gathered nothing.
-        """
-        deadline = self._timeout_deadline(timeout)
-        retry_prompt = prompt + "\n\n" + _STRUCTURED_RETRY
-        session_budget = self._capped_session_budget(budget)
-        try:
-            session = self._factory.build_workflow_session(
-                prompt=retry_prompt,
-                budget=session_budget,
-                tools=[capture_tool],
-                isolation=isolation,
-                label=label,
-                tool_choice=_named_tool_choice(capture_tool.name),
-                thinking=False,
-            )
-        except Exception as exc:  # noqa: BLE001 — factory failure must not abort the fleet
-            await self.log(f"structured retry build failed ({label or 'agent'}): {exc}")
-            return None
-
-        self._track_session(session)
-        try:
-            self._carry_exploration(prior_session, session)
-            await self._run_session_turn(
-                session,
-                retry_prompt,
-                deadline=deadline,
-                cancel_event=capture_done,
-            )
-        except CallerTimeoutError:
-            await self.log(f"structured retry timed out ({label or 'agent'}) after {timeout}s")
-            return None
-        except Exception as exc:  # noqa: BLE001 — one dead agent never kills the fleet
-            await self.log(f"structured retry failed ({label or 'agent'}): {exc}")
-            return None
-        return capture_tool.captured if _schema_satisfied(capture_tool.captured, schema) else None
-
     async def _run_session_turn(
         self,
         session: Any,
@@ -1099,29 +583,6 @@ class WorkflowContext:
         )
         run_loop = session.run_loop() if cancel_event is None else session.run_loop(cancel_event)
         return await self._run_with_timeout(run_loop, self._remaining_timeout(deadline))
-
-    @staticmethod
-    def _carry_exploration(prior_session: Any, session: Any) -> None:
-        """Copy the first pass's conversation into the corrective session.
-
-        The corrective session is built fresh (seeded only with the system
-        prompt), so without this it would have none of the first pass's
-        exploration. We copy a *shallow list copy* of the prior messages — the
-        new list is independent (so the corrective turn's own appends don't
-        mutate the first session's history) while the message dicts are shared,
-        which is safe because neither side mutates a message in place.
-
-        Defensive: a session shape that lacks a settable ``messages`` (e.g. the
-        very first pass having failed before any message landed) must not abort
-        the corrective turn — the worst case is the pre-fix bare-prompt commit.
-        """
-        prior = getattr(prior_session, "messages", None)
-        if prior is None:
-            return
-        try:
-            session.messages = list(prior)
-        except Exception:  # noqa: BLE001 — carry-over is best-effort, never fatal
-            pass
 
     def _session_budget(self) -> int:
         lease = self._active_budget_lease.get()
